split indexer integration tests
